% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aloans}
\alias{aloans}
\title{Consumer automobile loans dataset}
\format{
\subsection{\code{aloans}}{

A data frame with 58,118 rows and 13 columns:
\describe{
\item{risk.cat}{Credit risk band based on the loan's Annual Percentage Rate
(APR). Categories include "super_prime" (0-5\%), "prime" (5-10\%), "near_prime"
(10-15\%), "subprime" (15-20\%), and "deep_subprime" (20\%+).}
\item{Z}{Time-to-event, representing the loan termination age in months.}
\item{Y}{Left-truncation time, representing the loan age in months
when the ABS trust began making payments to investors.}
\item{C}{Right-censoring indicator (1 = right-censored, 0 = exact
termination observed).}
\item{D}{Default indicator for competing risks (1 = default observed, 0 =
prepayment observed or right-censored).}
\item{R}{Prepayment indicator for competing risks (1 = prepayment observed,
0 = default observed or right-censored).}
\item{bond}{Categorical variable specifying the source ABS bond:
\code{"carmax"} (CarMax Auto Owner Trust 2017-2), \code{"aart"} (Ally Auto
Receivables Trust 2017-3), \code{"sdart"} (Santander Drive Auto Receivables
Trust 2017-2), or \code{"drive"} (Drive Auto Receivables Trust 2017-1).}
\item{orig.apr}{Obligor annual percentage rate}
\item{orig.term}{Original loan term (in months)}
\item{orig.loan.amt}{Original loan amount}
\item{abs0.age}{Loan age as of the start of the asset-backed security}
\item{abs0.balance}{Outstanding loan balance as of the start of the
asset-backed security}
\item{calc.pmt}{Calculated monthly payment}
}
}
}
\source{
Data was originally compiled from the Electronic Data Gathering,
Analysis, and Retrieval (EDGAR) system of the Securities and Exchange
Commission (SEC). The replication data repository is available at
\url{https://github.com/jackson-lautier/credit-risk-convergence/}.
}
\usage{
aloans
}
\description{
A dataset containing the performance records of 58,118 consumer automobile
loans with original terms of 72 to 73 months. The data was obtained from
four publicly traded asset-backed securities (ABS) bonds. This dataset is
subject to left-truncation (loans must survive long enough to be included in
the ABS trust) and right-censoring (the trust may be unwound before all
loans terminate), and it features competing risks of default and prepayment.
}
\references{
Lautier, J. P., Pozdnyakov, V., & Yan, J. (2024). On the
convergence of credit risk in current consumer automobile
loans. \emph{Journal of the Royal Statistical Society Series A: Statistics
in Society}, qnae137. \doi{10.1093/jrsssa/qnae137}.
}
\keyword{datasets}
