% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aloans_recov}
\alias{aloans_recov}
\title{Consumer automobile loans recovery curves}
\format{
\subsection{\code{aloans_recov}}{

A data frame with 365 rows and 3 columns:
\describe{
\item{risk.cat}{Credit risk band based on the loan's Annual Percentage Rate
(APR). Categories include "super_prime" (0-5\%), "prime" (5-10\%), "near_prime"
(10-15\%), "subprime" (15-20\%), and "deep_subprime" (20\%+).}
\item{month}{The month for which the recovery is estimated.}
\item{recovery}{The estimated recovery rate for a given month and credit risk band.}
}
}
}
\usage{
aloans_recov
}
\description{
A dataset containing the recovery curves associated with the automobile
loans analyzed in Lautier et al. 2024.
}
\references{
Lautier, J. P., Pozdnyakov, V., & Yan, J. (2024). On the
convergence of credit risk in current consumer automobile
loans. \emph{Journal of the Royal Statistical Society Series A: Statistics
in Society}, qnae137. \doi{10.1093/jrsssa/qnae137}.
}
\keyword{datasets}
