% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apv.R
\name{amort_bal}
\alias{amort_bal}
\title{Calculate Loan Amortization Schedule}
\usage{
amort_bal(time, orig_bal, int_rate, payment)
}
\arguments{
\item{time}{Vector of months}

\item{orig_bal}{Original loan balance}

\item{int_rate}{Monthly interest rate}

\item{payment}{Monthly payment}
}
\value{
Outstanding balance at each month
}
\description{
Calculate Loan Amortization Schedule
}
