% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-methods.R
\name{autoplot.alife}
\alias{autoplot.alife}
\title{Plot an 'alife' Object using ggplot2}
\usage{
autoplot.alife(object, ci_level = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{alife}.}

\item{ci_level}{A single number or numeric vector of confidence levels
to display (e.g., \code{0.95} or \code{c(0.80, 0.95)}).}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{ggplot} object.
}
\description{
Plot an 'alife' Object using ggplot2
}
