% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{calc_apv}
\alias{calc_apv}
\title{Calculate Actuarial Present Value (APV)}
\usage{
calc_apv(x, ...)
}
\arguments{
\item{x}{An object of class \code{acdf} or \code{alife}.}

\item{...}{Additional arguments.}
}
\value{
A numeric value representing the APV.
}
\description{
Calculate Actuarial Present Value (APV)
}
