% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apv.R
\name{calc_apv_var.acdf}
\alias{calc_apv_var.acdf}
\title{Calculate Variance of Actuarial Present Value}
\usage{
\method{calc_apv_var}{acdf}(
  x,
  payment,
  residual_value,
  elapsed_time,
  max_lifetime,
  max_delay,
  payment_cap,
  discount_factors,
  delay_prob_matrix,
  depreciation_matrix,
  ...
)
}
\arguments{
\item{x}{An object of class \code{acdf} or \code{alife}.}

\item{payment}{Monthly payment amount.}

\item{residual_value}{Contractual residual value.}

\item{elapsed_time}{Current age (time elapsed) of the lease.}

\item{max_lifetime}{Maximum time to consider (time horizon).}

\item{max_delay}{Maximum possible payment delay.}

\item{payment_cap}{Cap on the number of payments to be made.}

\item{discount_factors}{Numeric vector of monthly discount factors.}

\item{delay_prob_matrix}{Matrix where rows are termination times and columns are delay lengths.}

\item{depreciation_matrix}{Matrix with rows as termination times, first column is expectation and second column is variance of depreciation.}

\item{...}{Additional arguments.}
}
\description{
Computes the variance of the Actuarial Present Value (APV) of future
cash flows.
}
