% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{calc_apv_var}
\alias{calc_apv_var}
\title{Calculate Variance of APV}
\usage{
calc_apv_var(x, ...)
}
\arguments{
\item{x}{An object of class \code{acdf} or \code{alife}.}

\item{...}{Additional arguments.}
}
\value{
A numeric value representing the variance of the APV.
}
\description{
Calculate Variance of APV
}
