% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{calc_cdf}
\alias{calc_cdf}
\alias{calc_cdf.alife}
\alias{calc_cdf.alife_multi}
\title{Calculate CDF from Hazard Estimates}
\usage{
calc_cdf(x, ...)

\method{calc_cdf}{alife}(x, ...)

\method{calc_cdf}{alife_multi}(x, ...)
}
\arguments{
\item{x}{An object of class \code{alife} or \code{alife_multi}.}

\item{...}{Not used.}
}
\value{
An object of class \code{acdf} (columns \code{lifetime}, \code{cdf}, \code{se_cdf},
\code{density}, \code{se_dens}, and \code{risk_set}) or, if \code{x} is an \code{alife_multi}
object, of class \code{acdf_multi} (columns \code{event_type}, \code{lifetime},
\code{all_surv}, \code{cumulative}, \code{density}, and \code{risk_set}).
}
\description{
Returns an object of class \code{acdf}, corresponding to a
\code{data.frame} with tailored plot and summary methods. For competing risks
(\code{alife_multi} objects), the cause-specific cumulative incidence functions
are computed instead and an object of class \code{acdf_multi} is returned (see
also \code{\link[=calc_cif]{calc_cif()}}).
}
