% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{calc_osup}
\alias{calc_osup}
\title{Calculate Observed Support}
\usage{
calc_osup(lifetime, trunc_time)
}
\arguments{
\item{lifetime}{The vector of event or censoring times.}

\item{trunc_time}{The vector of left-truncation times.}
}
\value{
A numeric vector representing the observed support of the lifetime
variable. An empty vector (with a warning) if fewer than two time points
are available.
}
\description{
This helper function generates the observable support of the lifetime variable
based on the study's overall time range (Lautier et al. 2023,
\doi{10.1016/j.ecosta.2023.05.005}). In particular, it takes \eqn{\Delta}
as the smallest observed time, \code{min(c(lifetime, trunc_time))}, and
\eqn{\omega} as the largest observed lifetime, \code{max(lifetime)}, and
outputs a sequence ranging from \eqn{\Delta + 1} to \eqn{\omega} (see
\code{\link[=retrieve_support]{retrieve_support()}}).
}
