% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{calc_tp}
\alias{calc_tp}
\title{Calculate Default Time Points}
\usage{
calc_tp(time, trunc_time)
}
\arguments{
\item{time}{The vector of event or censoring times.}

\item{trunc_time}{The vector of left-truncation times.}
}
\value{
A numeric vector of default time points to evaluate the hazard at.
}
\description{
This helper function generates a default sequence of evaluation points based
on the study's overall time range (PUT REFERENCE HERE). In particular, it
calculates \eqn{\delta} and \eqn{m} based on left-truncation and
time-to-event variables and outputs a sequence ranging from \eqn{\delta + 1}
to \eqn{\delta + m}.
}
