% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apv.R
\name{disc_fac}
\alias{disc_fac}
\title{Calculate Discount Factors}
\usage{
disc_fac(time, int_rate)
}
\arguments{
\item{time}{Vector of months}

\item{int_rate}{Monthly discount rate}
}
\value{
Discount factor for each month
}
\description{
Calculate Discount Factors
}
