% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{.hazard_core}
\alias{.hazard_core}
\title{Hazard estimates over a support (auxiliary function for
\code{estimate_hazard})}
\usage{
.hazard_core(
  support,
  trunc_time,
  lifetime,
  censoring_indicator,
  event_indicator,
  carry_hazard
)
}
\arguments{
\item{support}{where to calculate the hazards}

\item{trunc_time}{A numeric vector representing the observed
left-truncation time. If \code{NULL} (the default), it is set to a vector of
0s (i.e., no left-truncation).}

\item{lifetime}{A numeric vector representing the observed time to
event (or censoring).}

\item{censoring_indicator}{An indicator for censoring (1 = censored, 0 =
not). Defaults to a vector of 0s if \code{NULL}. An observation is only
treated as an event (of a given type) if its censoring indicator is 0.}

\item{event_indicator}{A binary vector flagging the observations whose
event is of the type being estimated (all ones when there are no
competing risks).}

\item{carry_hazard}{A \code{logical} indicator on whether 0 hazard estimates
should be replaced by the last non-zero estimate. Defaults to \code{FALSE}.}
}
\value{
a \code{data.frame}
}
\description{
Hazard estimates over a support (auxiliary function for
\code{estimate_hazard})
}
\author{
lcgodoy
}
\keyword{internal}
