% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{f_hat}
\alias{f_hat}
\title{\eqn{\hat{f}(x)}}
\usage{
f_hat(t, lifetime, event, censoring_indicator)
}
\arguments{
\item{t}{A time point at which hazard estimates are sought.}

\item{lifetime}{A numeric vector representing the observed time to
event (or censoring).}

\item{event}{event indicator}

\item{censoring_indicator}{An indicator for censoring (1 = censored, 0 =
not). Defaults to a vector of 0s if \code{NULL}. An observation is only
treated as an event (of a given type) if its censoring indicator is 0.}
}
\value{
a scalar
}
\description{
\eqn{\hat{f}(x)}
}
\author{
lcgodoy
}
\keyword{internal}
