% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_0haz}
\alias{fix_0haz}
\title{Carry Forward Last Observation for Zero-Hazard Estimates (Internal Use)}
\usage{
fix_0haz(haz_est)
}
\arguments{
\item{haz_est}{A \code{data.frame} calculated using the function
\code{estimate_hazard}}
}
\value{
A \code{data.frame} with hazard estimates carried forward when zeros
are present.
}
\description{
Carry Forward Last Observation for Zero-Hazard Estimates (Internal Use)
}
\keyword{internal}
