% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-methods.R
\name{ggauto}
\alias{ggauto}
\title{Plot an \code{alife} (or \code{alife_multi}) Object using \code{ggplot2}}
\usage{
ggauto(object, ...)
}
\arguments{
\item{object}{An object of class \code{alife} or \code{alife_multi}.}

\item{...}{Arguments passed to methods (e.g., \code{ci_level}).}
}
\value{
A \code{ggplot} object.
}
\description{
Plot an \code{alife} (or \code{alife_multi}) Object using \code{ggplot2}
}
