% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{new_acdf}
\alias{new_acdf}
\title{Create an acdf Object}
\usage{
new_acdf(x = data.frame())
}
\arguments{
\item{x}{A \code{data.frame} that has the required columns for an \code{acdf}
object: \code{lifetime} and \code{density}. Defaults to an empty \code{data.frame}.}
}
\value{
An object of class \code{acdf}.
}
\description{
Constructs an object of class \code{acdf} from a data frame, after validating
its structure. Note that this function is mostly for internal usage.
}
\examples{
# Create a minimal data frame with the required columns
df <- data.frame(lifetime = 1:2, cdf = c(0.1, 1),
                 density = c(0.1, 0.9))

# Construct the acdf object
my_acdf_obj <- new_acdf(df)
class(my_acdf_obj)

}
