% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{new_acdf_multi}
\alias{new_acdf_multi}
\title{Create an acdf_multi Object}
\usage{
new_acdf_multi(x = data.frame())
}
\arguments{
\item{x}{A \code{data.frame} that has the required columns for an \code{acdf}
object: \code{lifetime} and \code{density}. Defaults to an empty \code{data.frame}.}
}
\value{
An object of class \code{acdf_multi}.
}
\description{
Competing-risks counterpart of \code{\link[=new_acdf]{new_acdf()}}. Internal use.
}
\keyword{internal}
