% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{new_acif}
\alias{new_acif}
\title{Create an acif Object}
\usage{
new_acif(x = data.frame())
}
\arguments{
\item{x}{A \code{data.frame} that has the required columns for an \code{acif}
object: \code{lifetime}, \code{event_type}, \code{all_surv}, and \code{cif}. Defaults to an
empty \code{data.frame}.}
}
\value{
An object of class \code{acif}.
}
\description{
Constructs an object of class \code{acif} from a data frame, after validating
its structure. Note that this function is mostly for internal usage.
}
