% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apv.R
\name{payment_pv}
\alias{payment_pv}
\title{Internal helper for payment present value}
\usage{
payment_pv(
  payment,
  lifetime,
  delay,
  elapsed_time,
  payment_cap,
  discount_factors
)
}
\arguments{
\item{payment}{Monthly payment.}

\item{lifetime}{Time of termination.}

\item{delay}{Delay in payment.}

\item{elapsed_time}{Current age.}

\item{payment_cap}{Cap on payments.}

\item{discount_factors}{Vector of discount factors.}
}
\description{
Internal helper for payment present value
}
\keyword{internal}
