% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.acdf}
\alias{plot.acdf}
\title{Plot Method for an 'acdf' Object}
\usage{
\method{plot}{acdf}(x, ci_level = 0.95, color = 2, col_line = 1, ...)
}
\arguments{
\item{x}{An object of class \code{acdf}. Typically the output of the
\code{calc_cdf} function.}

\item{ci_level}{A numeric vector of confidence levels to plot (e.g.,
\code{c(0.5, 0.95)}). Defaults to \code{0.95}.}

\item{color}{The color for the confidence interval polygon (CDF) and error
bars (density).}

\item{col_line}{The color for the CDF line and density points.}

\item{...}{Additional arguments passed to the base \code{plot} function (e.g.,
\code{main}, \code{xlab}, \code{ylab}, \code{ylim}).}
}
\value{
A plot of the CDF.
}
\description{
Plot Method for an 'acdf' Object
}
\seealso{
\code{\link[=calc_cdf]{calc_cdf()}}
}
