% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.alife_multi}
\alias{print.alife_multi}
\title{Print Method for an 'alife_multi' Object}
\usage{
\method{print}{alife_multi}(x, ...)
}
\arguments{
\item{x}{An object of class \code{alife_multi}. Typically the output of the
\code{estimate_hazard} function.}

\item{...}{Not used.}
}
\value{
Prints information about \code{x} and invisibly returns it.
}
\description{
Print Method for an 'alife_multi' Object
}
\seealso{
\code{\link[=estimate_hazard]{estimate_hazard()}}
}
