% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apv.R
\name{prob_delay}
\alias{prob_delay}
\title{Internal helper for probability of delay}
\usage{
prob_delay(delay, lifetime, elapsed_time, max_delay, delay_prob_matrix)
}
\arguments{
\item{delay}{Length of delay.}

\item{lifetime}{Time of termination.}

\item{elapsed_time}{Current age.}

\item{max_delay}{Maximum possible delay.}

\item{delay_prob_matrix}{Delay probability matrix.}
}
\description{
Internal helper for probability of delay
}
\keyword{internal}
