% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apv.R
\name{prob_termination}
\alias{prob_termination}
\title{Internal helper for probability of termination}
\usage{
prob_termination(x, lifetime, elapsed_time)
}
\arguments{
\item{x}{An \code{acdf} object.}

\item{lifetime}{Time of termination.}

\item{elapsed_time}{Current age.}
}
\description{
Internal helper for probability of termination
}
\keyword{internal}
