% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apv.R
\name{residual_pv}
\alias{residual_pv}
\title{Internal helper for residual present value}
\usage{
residual_pv(residual_value, lifetime, elapsed_time, discount_factors)
}
\arguments{
\item{residual_value}{Contractual residual value.}

\item{lifetime}{Time of termination.}

\item{elapsed_time}{Current age.}

\item{discount_factors}{Vector of discount factors.}
}
\description{
Internal helper for residual present value
}
\keyword{internal}
