% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aux_kmat}
\alias{aux_kmat}
\alias{rmat}
\alias{build_kmat}
\alias{build_rmat}
\alias{build_amat}
\alias{build_sigmat}
\alias{build_pmfvar}
\alias{build_cdfvar}
\title{Build the auxiliary matrix K}
\usage{
aux_kmat(hazard, rephaz, support_length)

build_kmat(hazard)

build_rmat(hazard)

build_amat(support_length)

build_sigmat(hazard, se_log_hazard)

build_pmfvar(hazard, se_log_hazard)

build_cdfvar(pmfvar)
}
\arguments{
\item{hazard}{A \code{vector} of estimated hazards at every timepoint.}

\item{rephaz}{number of times \code{hazard} must be repeated (auxiliary)}

\item{support_length}{length of \code{hazard}.}

\item{se_log_hazard}{A \code{vector} of SE estimates for the hazards at the
logit scale.}

\item{pmfvar}{a square \code{matrix} corresponding to the output of a
\code{build_pmfvar} call.}
}
\value{
A square \code{matrix} with number of rows (and columns) matching the
dimension of \code{hazard}.
}
\description{
Build the auxiliary matrix K
}
\author{
lcgodoy
}
\keyword{internal}
