% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.acdf}
\alias{summary.acdf}
\title{Summary Method for an 'acdf' Object}
\usage{
\method{summary}{acdf}(object, by = 5, ...)
}
\arguments{
\item{object}{An object of class \code{acdf}. Typically the output of the
\code{calc_cdf} method.}

\item{by}{an \code{integer} defining the periodicity of the summary.}

\item{...}{Not used.}
}
\value{
A \code{data.frame} with the CDF and density at every \code{by} time
points.
}
\description{
Summary Method for an 'acdf' Object
}
\seealso{
\code{\link[=calc_cdf]{calc_cdf()}}
}
