% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.acdf_multi}
\alias{summary.acdf_multi}
\title{Summary Method for an 'acdf_multi' Object}
\usage{
\method{summary}{acdf_multi}(object, by = 5, ...)
}
\arguments{
\item{object}{An object of class \code{acdf_multi}. Typically the output of the
\code{calc_cdf} method.}

\item{by}{an \code{integer} defining the periodicity of the summary.}

\item{...}{Not used.}
}
\value{
A \code{data.frame} with the all-cause survival, cause-specific
cumulative incidence, and density at every \code{by} time points.
}
\description{
Summary Method for an 'acdf_multi' Object
}
\seealso{
\code{\link[=calc_cdf]{calc_cdf()}}
}
