% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.acif}
\alias{summary.acif}
\title{Summary Method for an 'acif' Object}
\usage{
\method{summary}{acif}(object, by = 5, ...)
}
\arguments{
\item{object}{An object of class \code{acif}. Typically the output of the
\code{calc_cif} method.}

\item{by}{an \code{integer} defining the periodicity of the summary.}

\item{...}{Not used.}
}
\value{
A \code{data.frame} with the cause-specific CIFs and the all-cause
survival at every \code{by} time points.
}
\description{
Summary Method for an 'acif' Object
}
\seealso{
\code{\link[=calc_cif]{calc_cif()}}
}
