% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.acif_multi}
\alias{summary.acif_multi}
\title{Summary Method for an 'acif' Object}
\usage{
\method{summary}{acif_multi}(object, by = 5, ...)
}
\arguments{
\item{object}{An object of class \code{acif}. Typically the output of the
\code{calc_cif} method.}

\item{by}{an \code{integer} defining the periodicity of the summary.}

\item{...}{Additional arguments passed to the base \code{print} function (e.g.,
\code{digits}).}
}
\value{
A summary of the CIF function.
}
\description{
Summary Method for an 'acif' Object
}
\seealso{
\code{\link[=calc_cif]{calc_cif()}}
}
