% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.alife_multi}
\alias{summary.alife_multi}
\title{Summary Method for an 'alife_multi' Object}
\usage{
\method{summary}{alife_multi}(object, by = 5, ...)
}
\arguments{
\item{object}{An object of class \code{alife_multi}. Typically the output of
the \code{estimate_hazard} function.}

\item{by}{an \code{integer} defining the periodicity of the summary.}

\item{...}{Additional arguments passed to the base \code{print} function (e.g.,
\code{digits}).}
}
\value{
Prints the cause-specific hazard estimates at every \code{by} time
points and invisibly returns them as a \code{data.frame}.
}
\description{
Summary Method for an 'alife_multi' Object
}
\seealso{
\code{\link[=estimate_hazard]{estimate_hazard()}}
}
