% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{u_hat}
\alias{u_hat}
\title{\eqn{\hat{U}(x)}}
\usage{
u_hat(t, lifetime, trunc_time)
}
\arguments{
\item{t}{A time point at which hazard estimates are sought.}

\item{lifetime}{A numeric vector representing the observed time to
event (or censoring).}

\item{trunc_time}{A numeric vector representing the observed
left-truncation time. If \code{NULL} (the default), it is set to a vector of
0s (i.e., no left-truncation).}
}
\value{
a scalar
}
\description{
\eqn{\hat{U}(x)}
}
\author{
lcgodoy
}
\keyword{internal}
