% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{validate_acif}
\alias{validate_acif}
\title{Validate an acif Object}
\usage{
validate_acif(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
\code{NULL}, invisibly, if validation is successful. Throws an error on
failure.
}
\description{
Checks if an object is a valid \code{acif} object by ensuring it is a data
frame and contains all required columns.
}
\keyword{internal}
