% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{var_hat}
\alias{var_hat}
\title{Variance of the log-transformed hazard estimate}
\usage{
var_hat(lambda, risk_set, fh, n)
}
\arguments{
\item{lambda}{hazard rate (currently unused).}

\item{risk_set}{\eqn{\hat{U}}}

\item{fh}{\eqn{\hat{f}}}

\item{n}{sample size (or number of timepoints)}
}
\value{
a scalar
}
\description{
Variance of the log-transformed hazard estimate
}
\author{
lcgodoy
}
\keyword{internal}
