% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aart}
\alias{aart}
\title{Consumer automobile loans from Ally Auto Receivables Trust}
\format{
\subsection{\code{aart}}{

A data frame with 2,756 rows and 12 columns:
\describe{
\item{\code{Xi}}{Time-to-event, representing the loan termination age
in months.}
\item{\code{Yi}}{Left-truncation time, representing the loan age in
months when the ABS trust began making payments to investors.}
\item{\code{credit.score}}{Obligor credit score}
\item{\code{interest.rate}}{Obligor annual percentage rate}
\item{\code{pti}}{Obligor payment-to-income}
\item{\code{veh.value}}{Estimated vehicle value at signing (\eqn{\\log} scale)}
\item{\code{co.sign}}{Co-Obligor indicator (1 = \code{TRUE})}
\item{\code{new.used}}{New-Used indicator (1 = \code{NEW})}
\item{\code{subvent.rate}}{Subvented interest rate indicator (1 = \code{TRUE})}
\item{\code{subvent.cash}}{Cash rebate indicator (1 = \code{TRUE})\\}
\item{\code{veh.pick.up}}{Pick-up truck indicator (1 = \code{TRUE})}
\item{\code{veh.suv}}{Sport-Utility-Vehicle indicator (1 = \code{TRUE})}
}
}
}
\source{
Data was compiled from the Electronic Data Gathering, Analysis, and
Retrieval (EDGAR) system of the Securities and Exchange Commission
(SEC). The replication data repository is available at
\url{https://github.com/jackson-lautier/consumer-auto-abs-parametric}.
}
\usage{
aart
}
\description{
A dataset containing time-to-event and left-truncation times for consumer
automobile loans from the Ally Auto Receivables Trust (AART). The data were
compiled from the the Electronic Data Gathering, Analysis, and Retrieval
(EDGAR) system of the Securities and Exchange Comission (SEC).
}
\references{
Lautier, J. P., Pozdnyakov, V., & Yan, J. (2025). Estimating the
time-to-event distribution for loan-level data within a consumer auto loan
asset-backed security. \emph{The Annals of Applied Statistics}.
\doi{10.1214/25-AOAS2103}.
}
\keyword{datasets}
