% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aart}
\alias{aart}
\title{NAME OF THE DATASET?}
\format{
\subsection{\code{aart}}{

A data frame with 47,315 rows and 3 columns:
\describe{

\item{\code{risk_cat}}{Credit score category}
\item{\code{Zi}}{Time-of-event}
\item{\code{Yi}}{Truncation time}
\item{\code{Di}}{Default indicator (1 = default, 0 = not default)}
}
}
}
\source{
DATASETS SOURCE (I'll get from the paper)
}
\usage{
aart
}
\description{
We should have a brief description here
}
\keyword{datasets}
