% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aloans}
\alias{aloans}
\title{Consumer automobile loans dataset}
\format{
\subsection{\code{aloans}}{

A data frame with 58,118 rows and 7 columns:
\describe{
\item{risk_cat}{Risk category}
\item{Z}{Time-of-event}
\item{Y}{Truncation time}
\item{C}{Censor indicator (1 = censored, 0 = not censored)}
\item{D}{Default indicator (1 = default, 0 = not default)}
\item{bond}{Categorical variable specifying different bonds.}
}
}
}
\source{
DATASETS SOURCE (I'll get from the paper)
}
\usage{
aloans
}
\description{
We should have a brief description here
}
\keyword{datasets}
