% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{calc_cdf}
\alias{calc_cdf}
\alias{calc_cdf.alife}
\alias{calc_cdf.alife_multi}
\title{Calculate CDF from Hazard Estimates}
\usage{
calc_cdf(x, ...)

\method{calc_cdf}{alife}(x, ...)

\method{calc_cdf}{alife_multi}(x, ...)
}
\arguments{
\item{x}{An object of class \code{alife}}

\item{...}{Not used.}
}
\value{
An object of class \code{acdf}.
}
\description{
Returns an object of class \code{acdf}, corresponding to a
\code{data.frame} with taylored plot, summary and print methods.
}
