% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{calc_cdf}
\alias{calc_cdf}
\alias{calc_cdf.alife}
\alias{calc_cdf.alife_multi}
\title{Calculate CDF from Hazard Estimates}
\usage{
calc_cdf(x, ...)

\method{calc_cdf}{alife}(x, ...)

\method{calc_cdf}{alife_multi}(x, ...)
}
\arguments{
\item{x}{An object of class \code{alife} or \code{alife-multi}.}

\item{...}{Not used.}
}
\value{
The original object \code{x} with a new \code{cdf} column.
}
\description{
Adds a 'cdf' column to an 'alife' or 'alife-multi' object based
on the hazard estimates.
}
