% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{calc_cif}
\alias{calc_cif}
\alias{calc_cif.alife_multi}
\title{Calculate cumulative incidence functions (CIF) from Hazard estimates}
\usage{
calc_cif(x, ...)

\method{calc_cif}{alife_multi}(x, ...)
}
\description{
Calculate cumulative incidence functions (CIF) from Hazard estimates
}
