% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{calc_osup}
\alias{calc_osup}
\title{Calculate Observed Support}
\usage{
calc_osup(lifetime, trunc_time)
}
\arguments{
\item{lifetime}{The vector of event or censoring times.}

\item{trunc_time}{The vector of left-truncation times.}
}
\value{
A numeric vector representing the observed support of the lifetime
variable.
}
\description{
This helper function generates the observed support of the lifetime variable
based on the study's overall time range (Lautier et al. 2023, <DOI:
10.1016/j.ecosta.2023.05.005>). In particular, it calculates \eqn{\Delta}
and \eqn{m} based on left-truncation and time-to-event variables and outputs
a sequence ranging from \eqn{\Delta + 1} to \eqn{\xi}, where \eqn{\xi =
\min(\omega, \varepsilon - 1)}, with \eqn{\varepsilon} denoting the present
time.
}
