% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apv.R
\name{calculate_apv}
\alias{calculate_apv}
\title{Calculate Actuarial Present Value (APV) of a Loan}
\usage{
calculate_apv(
  x,
  cur_age,
  orig_term,
  orig_loan_amt,
  mon_pmt,
  ref_rate,
  recov_curve,
  orig_apy = 0.15,
  type = "constant"
)
}
\arguments{
\item{x}{An object of class \code{acdf}. Typically the output of the
\code{estimate_hazard} function.}

\item{cur_age}{Current age of the loan in months.}

\item{orig_term}{Original loan term in months.}

\item{orig_loan_amt}{Original loan amount.}

\item{mon_pmt}{Monthly payment amount.}

\item{ref_rate}{Annualized reference discount rate.}

\item{recov_curve}{Recovery curve data frame (columns: Month, Recovery).}

\item{orig_apy}{Internal loan APY for amortization. Defaults to 0.15.}

\item{type}{A character string specifying the interpolation type: either
\code{"constant"} (default), \code{"geometric"} or \code{"linear"}.}
}
\value{
A list containing APV, second moment (APV2), standard deviation
(SD).
}
\description{
Calculates the APV of a loan under competing risks (Default vs. Pre-payment)
and left-truncated survival data.
}
