% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{.hazard_core}
\alias{.hazard_core}
\title{Auxiliary function for \code{estimate_hazard}}
\usage{
.hazard_core(
  support,
  trunc_time,
  time_to_event,
  censoring,
  event_indicator,
  carry_hazard
)
}
\arguments{
\item{support}{where to calculate the hazards}

\item{trunc_time}{A numeric vector representing the observed left-truncated
time.}

\item{time_to_event}{A numeric vector representing the observed time to
event.}

\item{censoring}{An indicator for censoring (1=censored, 0=not). Defaults to
a vector of 0s if \code{NULL}. An observation is only treated as an event if
status=1 AND censoring=0.}

\item{event_indicator}{legacy.}

\item{carry_hazard}{A \code{boolean} indicator on whether 0 hazard estimates
should be replaced by the last non-zero estimate. Defaults to \code{FALSE}}
}
\value{
a \code{data.frame}
}
\description{
Auxiliary function for \code{estimate_hazard}
}
\author{
lcgodoy
}
