% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{estimate_hazard}
\alias{estimate_hazard}
\title{Hazard rate}
\usage{
estimate_hazard(
  lifetime,
  trunc_time = NULL,
  censoring_indicator = NULL,
  event_type = NULL,
  support_lifetime_rv = NULL,
  carry_hazard = FALSE,
  ci_level = 0.95
)
}
\arguments{
\item{lifetime}{A numeric vector representing the observed time to
event.}

\item{trunc_time}{A numeric vector representing the observed left-truncated
time.}

\item{censoring_indicator}{An indicator for censoring (1=censored, 0=not). Defaults to
a vector of 0s if \code{NULL}. An observation is only treated as an event if
status=1 AND censoring=0.}

\item{event_type}{a vector of "events identifies" (experimental)}

\item{support_lifetime_rv}{A \code{vector} of time points at which to evaluate
the hazard.  If \code{NULL} (the default), it is calculated for a sequence from
\code{Delta + 1} to \code{omega} (that is, \code{max(lifetime)}).}

\item{carry_hazard}{A \code{boolean} indicator on whether 0 hazard estimates
should be replaced by the last non-zero estimate. Defaults to \code{FALSE}}

\item{ci_level}{A number between 0 and 1 indicating the level of the
confidence intervals.}
}
\value{
A \code{data.frame} with the hazard estimate their standard errors and
asymptotic confidence intervals. Importantly, in the output, the column
called \code{se_log_hazard} represents the standard error of the \code{hazard} rate
at the \code{logit} scale.
}
\description{
Estimate the non-parametric hazard rate for truncated and
censored data
}
\details{
Point estimate and asymptotic confidence intervals are calculated
based on \if{html}{\out{<REFERENCES>}} (We can also include some brief notation/definitions
here)
}
