% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{estimate_hazard}
\alias{estimate_hazard}
\title{Hazard rate}
\usage{
estimate_hazard(
  time_to_event,
  trunc_time = NULL,
  event_type = NULL,
  censoring = NULL,
  support_event = NULL,
  return_cdf = TRUE,
  carry_hazard = FALSE
)
}
\arguments{
\item{time_to_event}{The vector of event or censoring times.}

\item{trunc_time}{The vector of left-truncation times. Defaults to NULL (no
truncation), which is equivalent to a vector of zeros.}

\item{event_type}{The event indicator vector (1=default,
0=censored). Defaults to NULL (no censoring), which is equivalent to a
vector of ones.}

\item{censoring}{An indicator for censoring (1=censored, 0=not). Defaults to
a vector of 0s if \code{NULL}. An observation is only treated as an event if
status=1 AND censoring=0.}

\item{support_event}{A \code{vector} of time points at which to evaluate the
hazard.  If \code{NULL} (the default), it is calculated for a sequence from
\code{delta + 1} to \verb{??}.}

\item{return_cdf}{A \code{boolean} indicator on whether to return the estimated
CDF associated to the hazard rate or not. Default is \code{TRUE}}

\item{carry_hazard}{A \code{boolean} indicator on whether 0 hazard estimates
should be replaced by the last non-zero estimate. Defaults to \code{FALSE}}
}
\value{
A \code{data.frame} with the hazard estimate their standard errors and
asymptotic confidence intervals.
}
\description{
Estimate the non-parametric hazard rate for truncated and
censored data
}
\details{
Point estimate and asymptotic confidence intervals are calculated
based on \if{html}{\out{<REFERENCES>}} (We can also include some brief notation/definitions
here)
}
