% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{estimate_hazard}
\alias{estimate_hazard}
\title{Hazard rate}
\usage{
estimate_hazard(
  time,
  trunc_time = NULL,
  status = NULL,
  censoring = NULL,
  eval_points = NULL,
  return_cdf = TRUE
)
}
\arguments{
\item{time}{The vector of event or censoring times.}

\item{trunc_time}{The vector of left-truncation times. Defaults to NULL (no
truncation), which is equivalent to a vector of zeros.}

\item{status}{The event indicator vector (1=event, 0=censored). Defaults to
NULL (no censoring), which is equivalent to a vector of ones.}

\item{censoring}{An indicator for censoring (1=censored, 0=not). Defaults to
a vector of 0s if \code{NULL}. An observation is only treated as an event if
status=1 AND censoring=0.}

\item{eval_points}{A \code{vector} of time points at which to evaluate the
hazard.  If \code{NULL} (the default), it is calculated for a sequence from
\code{delta + 1} to \code{delta + m}.}

\item{return_cdf}{A \code{boolean} indicator on whether to return the estimated
CDF associated to the hazard rate or not. Default is \code{TRUE}}
}
\value{
A \code{data.frame} with the hazard estimate their standard errors and
asymptotic confidence intervals.
}
\description{
Estimate the non-parametric hazard rate for truncated and
censored data
}
\details{
Point estimate and asymptotic confidence intervals are calculated
based on \if{html}{\out{<REFERENCES>}} (We can also include some brief notation/definitions
here)
}
