% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ev}
\alias{ev}
\alias{ev_life}
\alias{ev_life.acdf}
\alias{ev_life.acdf_multi}
\alias{ev_life.alife}
\title{Calculate expected value for the time to event of interest}
\usage{
ev_life(x, digits = 2, ...)

\method{ev_life}{acdf}(x, digits = 2, ...)

\method{ev_life}{acdf_multi}(x, digits = 2, ...)

\method{ev_life}{alife}(x, ...)
}
\arguments{
\item{x}{An object of class \code{acdf}.}

\item{digits}{number of digits for the output.}

\item{...}{extra arguments to be passed to the \code{round} function.}
}
\value{
The expected value associated with the object \code{x}.
}
\description{
Uses the estimated density from the 'acdf' output.
}
