% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{extend_hazard}
\alias{extend_hazard}
\alias{extend_hazard.alife}
\alias{extend_hazard.alife_multi}
\title{Extend Hazard Rates}
\usage{
extend_hazard(
  x,
  end = 72,
  type = c("constant", "geometric", "linear"),
  end_event = "Prepayment"
)

\method{extend_hazard}{alife}(
  x,
  end = 72,
  type = c("constant", "geometric", "linear"),
  end_event = "Prepayment"
)

\method{extend_hazard}{alife_multi}(
  x,
  end = 72,
  type = c("constant", "geometric", "linear"),
  end_event = "Prepayment"
)
}
\arguments{
\item{x}{An object of class \code{alife} or \code{alife_multi}.}

\item{end}{A numeric scalar indicating the target end time for the
support. Typically, the original loan term.}

\item{type}{A character string specifying the interpolation type: either
\code{"constant"}, \code{"geometric"} or \code{"linear"}.}

\item{end_event}{A string indicating the event that will happen \emph{certainly}
at \emph{end}. Only used when \code{x} is of class \code{alife_multi}.}
}
\value{
An object of the same class as \code{x} with the extended lifetime
support.
}
\description{
Extends the estimated hazard rates up to a specified end time
using linear or geometric interpolation. This is only used for the
\code{apv} related functions when the observed lifetime support exceeds
the original loan term.
}
