% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{f_hat}
\alias{f_hat}
\title{\eqn{\hat{f}(x)}}
\usage{
f_hat(t, time_to_event, event, censoring)
}
\arguments{
\item{t}{A time point at which hazard estimates are sought.}

\item{time_to_event}{A numeric vector representing the observed time to
event.}

\item{event}{event indicator}

\item{censoring}{An indicator for censoring (1=censored, 0=not). Defaults to
a vector of 0s if \code{NULL}. An observation is only treated as an event if
status=1 AND censoring=0.}
}
\value{
a scalar
}
\description{
\eqn{\hat{f}(x)}
}
\author{
lcgodoy
}
