% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-methods.R
\name{ggauto.alife_multi}
\alias{ggauto.alife_multi}
\title{Plot an 'alife-multi' Object using \code{ggplot2}}
\usage{
\method{ggauto}{alife_multi}(object, ci_level = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{alife-multi}.}

\item{ci_level}{A single number or numeric vector of confidence levels
to display (e.g., \code{0.95} or \code{c(0.80, 0.95)}).}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{ggplot} object.
}
\description{
Creates a faceted plot, with one panel per event type.
}
