% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-methods.R
\name{ggauto}
\alias{ggauto}
\title{Plot an 'alife' (or \code{multi_alife}) Object using \code{ggplot2}}
\usage{
ggauto(object, ...)
}
\arguments{
\item{object}{An object of class \code{alife} or \code{alife-multi}.}

\item{...}{Not used.}
}
\value{
ggplot.
}
\description{
Plot an 'alife' (or \code{multi_alife}) Object using \code{ggplot2}
}
