% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lt_pmf}
\alias{lt_pmf}
\alias{.lt_core}
\title{Core for left-truncation PMF computation}
\usage{
.lt_core(lifetime, denom, trunc_time, rc = TRUE)

lt_pmf(x, trunc_time, rc = TRUE)
}
\arguments{
\item{lifetime}{A numeric vector representing the observed time to
event.}

\item{denom}{A \code{numeric} vector representing the denominator for the
calculation of the pmf}

\item{trunc_time}{A numeric vector representing the observed left-truncated
time.}

\item{rc}{a \code{numeric} boolean indicating whether right-censoring is to
be considered.}

\item{x}{is the output of an \code{estimate_hazard} call.}
}
\value{
A \code{numeric} vector of estimates of the PMF for the
left-truncation random variable.
}
\description{
Internal use
}
\author{
lcgodoy
}
