% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mbalt}
\alias{mbalt}
\title{NAME OF THE DATASET?}
\format{
\subsection{\code{mbalt}}{

A data frame with 1,553 rows and 2 columns:
\describe{

\item{\code{risk_cat}}{Credit score category}
\item{\code{Xi}}{Time-of-event}
\item{\code{Yi}}{Truncation time}
}
}
}
\source{
DATASETS SOURCE (I'll get from the paper)
}
\usage{
mbalt
}
\description{
We should have a brief description here
}
\keyword{datasets}
