% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mci}
\alias{mci}
\alias{multiple_cis}
\alias{multiple_cis.alife}
\alias{multiple_cis.alife_multi}
\title{Computing multiple CIs at once.}
\usage{
multiple_cis(x, ci_level = 0.95)

\method{multiple_cis}{alife}(x, ci_level = 0.95)

\method{multiple_cis}{alife_multi}(x, ci_level = 0.95)
}
\arguments{
\item{x}{An object of class \code{alife}. Typically the output of the
\code{estimate_hazard} function.}

\item{ci_level}{A numeric vector of confidence ci_level to plot (e.g.,
\code{c(0.5, 0.95)}). Defaults to \code{0.95}.}
}
\value{
a \code{list}
}
\description{
Computing multiple CIs at once.
}
\author{
lcgodoy
}
