% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.acdf}
\alias{plot.acdf}
\title{Plot Method for an 'acdf' Object}
\usage{
\method{plot}{acdf}(x, ...)
}
\arguments{
\item{x}{An object of class \code{acdf}. Typically the output of the
\code{estimate_hazard} function.}

\item{...}{Additional arguments passed to the base \code{plot} function (e.g.,
\code{xlab}, \code{ylab}, \code{ylim}).}
}
\value{
A plot of the CDF.
}
\description{
Plot Method for an 'acdf' Object
}
\seealso{
\code{\link[=calc_cdf]{calc_cdf()}}
}
