% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.acdf_multi}
\alias{plot.acdf_multi}
\title{Plot Method for an 'acdf_multi' Object}
\usage{
\method{plot}{acdf_multi}(x, ...)
}
\arguments{
\item{x}{An object of class \code{acdf_multi}.}

\item{...}{Additional arguments passed to the base \code{plot} function (e.g.,
\code{xlab}, \code{ylab}, \code{ylim}).}
}
\value{
A faceted plot of the CDFs.
}
\description{
Creates a faceted plot, with one panel per event type.
}
\seealso{
\code{\link[=calc_cdf]{calc_cdf()}}
}
