% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.alife}
\alias{plot.alife}
\title{Plot Method for an 'alife' Object}
\usage{
\method{plot}{alife}(x, ci_level = 0.95, color = 2, col_line = 1, ...)
}
\arguments{
\item{x}{An object of class \code{alife}. Typically the output of the
\code{estimate_hazard} function.}

\item{ci_level}{A numeric vector of confidence ci_level to plot (e.g.,
\code{c(0.5, 0.95)}). Defaults to \code{0.95}.}

\item{color}{The color for the confidence interval polygon.}

\item{col_line}{The color for the hazard rate line.}

\item{...}{Additional arguments passed to the base \code{plot} function (e.g.,
\code{main}, \code{xlab}, \code{ylab}, \code{ylim}).}
}
\value{
A plot of the hazard rate.
}
\description{
Plot Method for an 'alife' Object
}
\seealso{
\code{\link[=estimate_hazard]{estimate_hazard()}}
}
