% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.alife}
\alias{print.alife}
\title{Print Method for an 'alife' Object}
\usage{
\method{print}{alife}(x, ...)
}
\arguments{
\item{x}{An object of class \code{alife}. Typically the output of the
\code{estimate_hazard} function.}

\item{...}{Additional arguments passed to the base \code{print} function (e.g.,
\code{digits}).}
}
\value{
Prints information about \code{x} and invisibly returns it.
}
\description{
Print Method for an 'alife' Object
}
\seealso{
\code{\link[=estimate_hazard]{estimate_hazard()}}
}
