% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.alife_multi}
\alias{print.alife_multi}
\title{Print Method for an 'alife_multi' Object}
\usage{
\method{print}{alife_multi}(x, ...)

\method{print}{alife_multi}(x, ...)
}
\arguments{
\item{x}{An object of class \code{alife_multi}.}

\item{...}{Additional arguments passed to the base \code{print} function.}
}
\value{
Prints information about \code{x} and invisibly returns it.

Prints information about \code{x} and invisibly returns it.
}
\description{
Print Method for an 'alife_multi' Object

Print Method for an 'alife_multi' Object
}
\seealso{
\code{\link[=estimate_hazard]{estimate_hazard()}}
}
