% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ralife_cdf}
\alias{ralife_cdf}
\title{Samples from a "abslife" object}
\usage{
ralife_cdf(n, x)
}
\arguments{
\item{n}{number of samples}

\item{x}{the output of a \code{calc_cdf} function}
}
\value{
a vector of "time to event" samples
}
\description{
Samples from a "abslife" object
}
\author{
lcgodoy
}
