% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{retrieve_support}
\alias{retrieve_support}
\title{Computes the theoretical support of a lifetime variable}
\usage{
retrieve_support(Delta, m, omega, epsilon = NULL)
}
\arguments{
\item{Delta}{a scalar denoting the marketing period (in months) during which
the ABS is marketed to prospective investors after the trust closes to new
loan originations.}

\item{m}{a scalar representing the total number of months over which the
underlying consumer auto loans are originated. This defines the length of
the loan origination window before the trust closes.}

\item{omega}{a scalar denoting the known, finite upper bound of a loan's
lifetime (in months), which is fixed by its amortization schedule at
contract signing (for example, \eqn{\omega = 72} for a 72-month loan).}

\item{epsilon}{a scalar (default is \code{NULL}) denoting denotes the
present time (the current calendar month of observation) of the data
generation process for an active ABS pool. If larger than \code{omega},
right-censoring is present because it indicates some active loans are
still making ongoing payments and have not yet terminated.}
}
\value{
A numeric vector representing the theoretical or observed support
associated with the inputs.
}
\description{
This helper function generates a sequence of evaluation points (Lautier et
al. 2025, \url{DOI:10.1214/25-AOAS2103}).
}
\note{
The observed support may not extend all the way to \eqn{omega}
(\code{omega}). In that scenario, one may want to set \eqn{omega} to the
age of the oldest contract in the bond (corresponding to \eqn{\min(\omega,
  \varepsilon - 1)} in the notation used in Lautier et al. 2025), with
\eqn{\varepsilon} denoting the present time.
}
