% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{single_t_hazard}
\alias{single_t_hazard}
\title{Hazard estimate for a single time-point.}
\usage{
single_t_hazard(t, trunc_time, time_to_event, event)
}
\arguments{
\item{t}{A time point at which hazard estimates are sought.}

\item{trunc_time}{A numeric vector representing the observed left-truncated
time.}

\item{time_to_event}{A numeric vector representing the observed time to
event.}

\item{event}{The event indicator vector (1=default, 0=censored). Defaults to
NULL (no censoring), which is equivalent to a vector of ones.}
}
\value{
A vector containing the time to event, \eqn{\hat{C}_n}, the number
of events, and the hazard estimate along with its standard error.
}
\description{
Internal use.
}
\author{
lcgodoy
}
