% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{single_t_hazard}
\alias{single_t_hazard}
\title{Hazard estimate for a single time-point.}
\usage{
single_t_hazard(t, trunc_time, lifetime, event, censoring_indicator)
}
\arguments{
\item{t}{A time point at which hazard estimates are sought.}

\item{trunc_time}{A numeric vector representing the observed left-truncated
time.}

\item{lifetime}{A numeric vector representing the observed time to
event.}

\item{event}{event indicator}

\item{censoring_indicator}{An indicator for censoring (1=censored, 0=not). Defaults to
a vector of 0s if \code{NULL}. An observation is only treated as an event if
status=1 AND censoring=0.}
}
\value{
A vector containing the time to event, \eqn{\hat{C}_n}, the number
of events, and the hazard estimate along with its standard error.
}
\description{
Internal use.
}
\author{
lcgodoy
}
