% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apv.R
\name{solve_irr}
\alias{solve_irr}
\title{Solve for Risk-Adjusted Internal Rate of Return (IRR)}
\usage{
solve_irr(
  x,
  abs0_bal,
  cur_age,
  orig_term,
  orig_loan_amt,
  mon_pmt,
  recov_curve,
  orig_apy = 0.15,
  type = "constant"
)
}
\arguments{
\item{x}{An object of class \code{acdf}. Typically the output of the
\code{estimate_hazard} function.}

\item{abs0_bal}{Current loan balance (to match EPV against).}

\item{cur_age}{Current age of the loan in months.}

\item{orig_term}{Original loan term in months.}

\item{orig_loan_amt}{Original loan amount.}

\item{mon_pmt}{Monthly payment amount.}

\item{recov_curve}{Recovery curve data frame.}

\item{orig_apy}{Internal loan APY for amortization. Defaults to 0.15.}

\item{type}{A character string specifying the interpolation type: either
\code{"constant"} (default), \code{"geometric"} or \code{"linear"}.}
}
\value{
List with risk adjusted monthly rate, annualized rate, and objective
value.
}
\description{
Solve for Risk-Adjusted Internal Rate of Return (IRR)
}
