% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.acdf_multi}
\alias{summary.acdf_multi}
\title{Summary Method for an 'alife_multi' Object}
\usage{
\method{summary}{acdf_multi}(object, by = 5, ...)
}
\arguments{
\item{object}{An object of class \code{acdf_multi}. Typically the output of the
\code{calc_cdf} method.}

\item{by}{an \code{integer} defining the periodicity of the summary.}

\item{...}{Additional arguments passed to the base \code{print} function (e.g.,
\code{digits}).}
}
\value{
A summary of the hazard rate.
}
\description{
Summary Method for an 'alife_multi' Object
}
\seealso{
\code{\link[=estimate_hazard]{estimate_hazard()}}
}
