% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.alife}
\alias{summary.alife}
\title{Summary Method for an 'alife' Object}
\usage{
\method{summary}{alife}(object, by = 5, ...)
}
\arguments{
\item{object}{An object of class \code{alife}. Typically the output of the
\code{estimate_hazard} function.}

\item{by}{an \code{integer} defining the periodicity of the summary.}

\item{...}{Additional arguments passed to the base \code{print} function (e.g.,
\code{digits}).}
}
\value{
A summary of the hazard rate.
}
\description{
Summary Method for an 'alife' Object
}
\seealso{
\code{\link[=estimate_hazard]{estimate_hazard()}}
}
