% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{u_hat}
\alias{u_hat}
\title{\eqn{\hat{U}(x)}}
\usage{
u_hat(t, lifetime, trunc_time)
}
\arguments{
\item{t}{A time point at which hazard estimates are sought.}

\item{lifetime}{A numeric vector representing the observed time to
event.}

\item{trunc_time}{A numeric vector representing the observed left-truncated
time.}
}
\value{
a scalar
}
\description{
\eqn{\hat{U}(x)}
}
\author{
lcgodoy
}
