% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{validate_alife}
\alias{validate_alife}
\title{Validate an alife Object}
\usage{
validate_alife(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
The input \code{x}, invisibly, if validation is successful. Throws an
error on failure.
}
\description{
Checks if an object is a valid \code{alife} object by ensuring it is a data
frame and contains all required columns.
}
\keyword{internal}
