% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{var_hat}
\alias{var_hat}
\title{Variance of the log-transformed hazard estimate}
\usage{
var_hat(lambda, uh, fh, n)
}
\arguments{
\item{lambda}{hazard rate}

\item{uh}{\eqn{hat{U}}}

\item{fh}{\eqn{hat{f}}}
}
\value{
a scalar
}
\description{
Variance of the log-transformed hazard estimate
}
\author{
lcgodoy
}
